+++ tests/cpp/testobject/InvalidPdxUsage.cpp
@@ -837,7 +837,7 @@ bool InvalidPdxUsage::equals(PdxTests::InvalidPdxUsage& other,
   // lengthArr);
 
   if (!isPdxReadSerialized) {
-    for (int i = 0; i < m_objectArray->size(); i++) {
+    for (size_t i = 0; i < m_objectArray->size(); i++) {
       AddressWithInvalidAPIUsage* otherAddr1 =
           dynamic_cast<AddressWithInvalidAPIUsage*>(
               ot->m_objectArray->at(i).get());
@@ -855,12 +855,12 @@ bool InvalidPdxUsage::equals(PdxTests::InvalidPdxUsage& other,
   if (myenum->getEnumName() != otenum->getEnumName()) return false;
 
   genericValCompare(ot->m_arraylist->size(), m_arraylist->size());
-  for (int k = 0; k < m_arraylist->size(); k++) {
+  for (size_t k = 0; k < m_arraylist->size(); k++) {
     genericValCompare(ot->m_arraylist->at(k), m_arraylist->at(k));
   }
 
   genericValCompare(ot->m_vector->size(), m_vector->size());
-  for (int j = 0; j < m_vector->size(); j++) {
+  for (size_t j = 0; j < m_vector->size(); j++) {
     genericValCompare(ot->m_vector->at(j), m_vector->at(j));
   }
